+++ lib/cuke_linter.rb
@@ -8,6 +8,7 @@ require 'cuke_linter/linters/background_does_more_than_setup_linter'
 require 'cuke_linter/linters/element_with_too_many_tags_linter'
 require 'cuke_linter/linters/example_without_name_linter'
 require 'cuke_linter/linters/feature_file_with_invalid_name_linter'
+require 'cuke_linter/linters/feature_file_with_mismatched_name_linter'
 require 'cuke_linter/linters/feature_with_too_many_different_tags_linter'
 require 'cuke_linter/linters/feature_without_name_linter'
 require 'cuke_linter/linters/feature_without_description_linter'
@@ -36,6 +37,7 @@ module CukeLinter
                         'ElementWithTooManyTagsLinter'                 => ElementWithTooManyTagsLinter.new,
                         'ExampleWithoutNameLinter'                     => ExampleWithoutNameLinter.new,
                         'FeatureFileWithInvalidNameLinter'             => FeatureFileWithInvalidNameLinter.new,
+                        'FeatureFileWithMismatchedNameLinter'          => FeatureFileWithMismatchedNameLinter.new,
                         'FeatureWithTooManyDifferentTagsLinter'        => FeatureWithTooManyDifferentTagsLinter.new,
                         'FeatureWithoutDescriptionLinter'              => FeatureWithoutDescriptionLinter.new,
                         'FeatureWithoutNameLinter'                     => FeatureWithoutNameLinter.new,