+++ lib/cuke_linter.rb
@@ -5,6 +5,8 @@ require "cuke_linter/version"
 require 'cuke_linter/formatters/pretty_formatter'
 require 'cuke_linter/linters/linter'
 require 'cuke_linter/linters/background_does_more_than_setup_linter'
+require 'cuke_linter/linters/element_with_common_tags_linter'
+require 'cuke_linter/linters/element_with_duplicate_tags_linter'
 require 'cuke_linter/linters/element_with_too_many_tags_linter'
 require 'cuke_linter/linters/example_without_name_linter'
 require 'cuke_linter/linters/feature_file_with_invalid_name_linter'
@@ -34,6 +36,8 @@ require 'cuke_linter/linters/test_with_too_many_steps_linter'
 module CukeLinter
 
   @original_linters = { 'BackgroundDoesMoreThanSetupLinter'            => BackgroundDoesMoreThanSetupLinter.new,
+                        'ElementWithCommonTagsLinter'                  => ElementWithCommonTagsLinter.new,
+                        'ElementWithDuplicateTagsLinter'               => ElementWithDuplicateTagsLinter.new,
                         'ElementWithTooManyTagsLinter'                 => ElementWithTooManyTagsLinter.new,
                         'ExampleWithoutNameLinter'                     => ExampleWithoutNameLinter.new,
                         'FeatureFileWithInvalidNameLinter'             => FeatureFileWithInvalidNameLinter.new,