+++ lib/cuke_linter/linters/feature_file_with_mismatched_name_linter.rb
@@ -1,6 +1,6 @@
 module CukeLinter
 
-  # A linter that detects invalid feature file names
+  # A linter that detects mismatched feature file names
 
   class FeatureFileWithMismatchedNameLinter < Linter